example/scim/api-gateway/requirements.txt
Mako==1.3.10
pyzmq==26.4.0
Jinja2==3.1.6
scipy==1.14.1
matplotlib
opentelemetry-distro
opentelemetry-exporter-otlp-proto-http
